+++ util/misc.js
@@ -1,5 +1,40 @@
-define([], function(){
-	// This module defines miscellaneous utility methods.
+define(["put-selector/put"], function(put){
+	// summary:
+	//		This module defines miscellaneous utility methods for purposes of
+	//		adding styles, and throttling/debouncing function calls.
+	
+	// establish an extra stylesheet which addCssRule calls will use,
+	// plus an array to track actual indices in stylesheet for removal
+	var
+		extraSheet = put(document.getElementsByTagName("head")[0], "style"),
+		extraRules = [],
+		removeMethod;
+	
+	// Keep reference to actual StyleSheet object (.styleSheet for IE < 9)
+	extraSheet = extraSheet.sheet || extraSheet.styleSheet;
+	// Store name of method used to remove rules (removeRule for IE < 9)
+	removeMethod = extraSheet.deleteRule ? "deleteRule" : "removeRule";
+	
+	function removeRule(index){
+		// Function called by the remove method on objects returned by addCssRule.
+		var realIndex = extraRules[index],
+			i, l;
+		if (realIndex === undefined) { return; } // already removed
+		
+		// remove rule indicated in internal array at index
+		extraSheet[removeMethod](realIndex);
+		
+		// Clear internal array item representing rule that was just deleted.
+		// NOTE: we do NOT splice, since the point of this array is specifically
+		// to negotiate the splicing that occurs in the stylesheet itself!
+		extraRules[index] = undefined;
+		
+		// Then update array items as necessary to downshift remaining rule indices.
+		// Can start at index + 1, since array is sparse but strictly increasing.
+		for(i = index + 1, l = extraRules.length; i < l; i++){
+			if(extraRules[i] > realIndex){ extraRules[i]--; }
+		}
+	}
 	
 	var util = {
 		defaultDelay: 15,
@@ -14,7 +49,7 @@ define([], function(){
 				ran = true;
 				cb.apply(context, arguments);
 				setTimeout(function(){ ran = false; }, delay);
-			}
+			};
 		},
 		throttleDelayed: function(cb, context, delay){
 			// summary:
@@ -30,7 +65,7 @@ define([], function(){
 					ran = false;
 					cb.apply(context, a);
 				}, delay);
-			}
+			};
 		},
 		debounce: function(cb, context, delay){
 			// summary:
@@ -47,7 +82,22 @@ define([], function(){
 				timer = setTimeout(function(){
 					cb.apply(context, a);
 				}, delay);
-			}
+			};
+		},
+		
+		addCssRule: function(selector, css){
+			// summary:
+			//		Dynamically adds a style rule to the document.  Returns an object
+			//		with a remove method which can be called to later remove the rule.
+			
+			var index = extraRules.length;
+			extraRules[index] = (extraSheet.cssRules || extraSheet.rules).length;
+			extraSheet.addRule ?
+				extraSheet.addRule(selector, css) :
+				extraSheet.insertRule(selector + '{' + css + '}', extraRules[index]);
+			return {
+				remove: function(){ removeRule(index); }
+			};
 		}
 	};
 	return util;